feat : 게시글 삭제 기능 수정
1. 데이터 삭제 상태를 변경하는 방식
2. 댓글 연동 삭제
- 다른 사용자의 댓글이 존재하는 경우 삭제 불

+++ src/main/java/codesquad/springcafe/controller/ArticleController.java
@@ -79,9 +79,15 @@ public class ArticleController {
     @DeleteMapping("articles/{id}")
     public String delete(@PathVariable Long id, HttpSession session) {
         ArticleInfoDTO article = articleService.findById(id);
+        List<ReplyInfoDTO> replies = articleService.findRepliesById(id);
         UserInfoDTO user = (UserInfoDTO) session.getAttribute("loggedInUser");
+
         if (!article.isWriter(user.getUserId())) {
-            return "/article/delete_failed";
+            return "/article/delete_failed_user";
+        }
+
+        if (!replies.isEmpty() && !replies.stream().allMatch(reply -> reply.isWriter(user.getUserId()))) {
+            return "/article/delete_failed_reply";
         }
         articleService.delete(id);
         return "redirect:/";
